refactor: 웹팩 최적화 (#481)
* chore: install bundle-analyzer plugin (#479)

* refactor: code splitting (#479)

* chore: Tree-shaking sideEffects 설정 (#479)

* fix: msw 관련 코드 tree-shaking 되지 않는 오류 개선 (#479)

* style: msw handler 파일 및 폴더 구조 수정 (#479)

* docs: 셀럽잇 title 수정

* chore: 번들명 hashing 설정 (#479)
