+++ frontend/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "main": "index.tsx",
   "license": "MIT",
+  "sideEffects": false,
   "scripts": {
     "start:msw": "webpack serve --port 3000 --env TARGET_ENV=msw",
     "start:dev": "webpack serve --port 3000 --env TARGET_ENV=dev",
@@ -93,7 +94,8 @@
     "stylelint-config-standard": "^34.0.0",
     "stylelint-order": "^6.0.3",
     "terser-webpack-plugin": "^5.3.9",
-    "ts-jest": "^29.1.1"
+    "ts-jest": "^29.1.1",
+    "webpack-bundle-analyzer": "^4.9.1"
   },
   "eslintConfig": {
     "extends": [
