+++ frontend/public/index.html
@@ -20,7 +20,7 @@
     <meta property="og:url" content="celuveat.com" />
     <meta name="image" property="og:image" content="https://www.celuveat.com/og-image.jpeg" />
     <meta name="description" property="og:description" content="셀럽 추천 맛집 서비스, 셀럽잇" />
-    <title>celuveat</title>
+    <title>셀럽잇 - CELUVEAT</title>
   </head>
   <body>
     <noscript>
