+++ frontend/src/index.tsx
@@ -5,14 +5,9 @@ import App from '~/App';
 import GlobalStyles from './styles/GlobalStyles';
 import { worker } from './mocks/browser';
 
-const DEV = process.env.NODE_ENV === 'development';
-
-if (DEV) {
-  worker.start();
-}
+if (process.env.NODE_ENV === 'development') worker.start();
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
-
 const queryClient = new QueryClient();
 
 root.render(
@@ -20,7 +15,7 @@ root.render(
     <GlobalStyles />
     <QueryClientProvider client={queryClient}>
       <App />
-      {DEV && <ReactQueryDevtools />}
+      {process.env.NODE_ENV === 'development' && <ReactQueryDevtools />}
     </QueryClientProvider>
   </>,
 );
