+++ frontend/src/mocks/handler/index.ts
@@ -1,0 +1,3 @@
+export { DetailPageSuccessHandler } from '~/mocks/handler/detailPage/handler';
+export { MainPageSuccessHandler } from '~/mocks/handler/mainPage/handler';
+export { WishListPageSuccessHandler } from '~/mocks/handler/wishListPage/handler';
